Adjust Samplerate type to work without unconditional allocation
Adjust the Samplerate type, as used in any context where streamed audio
is being resampled, to doing so without unconditionally allocating a new
vector, but by using a provided output buffer instead. The process()
method will return how many samples of the input have been consumed and
how many output samples been produced in the process.

Change the method receiver to &mut self to make it obvious that
samplerate related state is being modified.

Also, remove a bunch of unnecessary getters and setters, which don't
really make much sense to have and/or may be actively broken (such as
the support for changing samplerates without making sure that the
resulting ratio is supported).

Lastly, while at it, rearrange the argument order of Samplerate::new()
to be more intuitive.