Remove public sub-modules
There is no good reason for us to expose the internal structure of the
crate in the public interface: the few types and functions we have are
best organized in a flat structure and that's what we already do. As
such, just make the two modules comprising the crate private.